requirements.txt
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
decorator==4.4.2
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
lxml==4.6.3
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
PyJWT==1.7.1
pytz==2021.1
selenium==3.141.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
